WIP: refactor module structure, reducing no of moduls
