+++ algebra.py
@@ -1,5 +1,85 @@
-from properties import *
+"""Generator functions of elementary matrices."""
+
 from itertools import chain
+from fractions import Fraction
+
+# Custom data types:
+
+#Supported fields are now \QQ and \RR
+#TODO: Add an inclusion of \ZZ into \QQ so that it's mathematically correct
+F = Fraction | float | int 
+
+Row = list[F]
+
+# Matrix is a list of list of ints. This is a list of rows.
+M = list[Row]
+
+### Elementary Matrices
+
+def I(n: int) -> M:
+    """return an identity matrix with dimension n"""
+
+    return [[0 if j != i else 1 for j in range(n)] for i in range(n)]
+
+def S(n: int, r1: int, r2: int) -> M:
+    """return a swap matrix by row r1 and r2 with dimension n"""
+
+    s = I(n)
+    s[r2], s[r1] = s[r1], s[r2]
+    return s
+
+def M(n: int, r1: int, a: F)  -> M:
+    """return a scale matrix by row r1 with argument a and dimension n"""
+
+    assert a < 0 or a > 0
+    elementary_scaled = I(n)
+    elementary_scaled[r1][r1] = a
+    return elementary_scaled
+
+def A(n: int, r1: int, r2: int, a: F) -> M:
+    """return an append matrix by row r1 and r2 with argument a and dimension n
+    
+    r1 = r1 + r2 * a
+    """
+
+    assert a < 0 or a > 0
+    elementary_added = I(n)
+    elementary_added[r1][r2] = 1 + a if r1 == r2 else a
+    return elementary_added
+
+### Properties
+
+def is_nullrow(row: Row) -> bool:
+    for value in row:
+        if value != 0:
+            return False
+    return True
+
+# duplicate to not introduce circular imports...
+def I(n: int) -> M:
+    """return an identity matrix with dimension n"""
+    return [[0 if j != i else 1 for j in range(n)] for i in range(n)]
+
+
+def is_identity_matrix(m: M) -> bool:
+    """Tests if matrix m is identity matrix"""
+    return m == I(len(m))
+
+
+def show(m: M):
+    isfloat = isinstance(m[0][0], float)
+    print('\n'.join(['\t'.join([f"{ele:.2f}" if isfloat else str(ele) for ele in row]) for row in m]))
+
+    # 2 digits after floating point, only works for floating point though!
+    #print('\n'.join(['\t'.join([f"{ele:.2f}" for ele in row]) for row in m]))
+    print()
+
+def show_ident(m: M, indent: int):
+    indentation = '\t' * indent
+    isfloat = isinstance(m[0][0], float)
+    for row in m:
+        print(indentation, end='')
+        print('\t'.join([f"{elem:.2f}" if isfloat else str(elem) for elem in row]))
 
 def get_pivot(row: Row) -> tuple[int | None, F | None]:
     # TODO: this doc-string is not uptodate
@@ -54,50 +134,6 @@ def mult(M1: M, M2: M) -> M:
     ]
 
 
-def blocking(M1: M, b: int) -> tuple[M, M, M, M]:
-    return (
-        [n[:b] for n in M1[:b]],
-        [n[b:] for n in M1[:b]],
-        [n[:b] for n in M1[b:]],
-        [n[b:] for n in M1[b:]],
-    )
-
-
-def mult_strassen(M1: M, M2: M) -> M:
-    """matrix multiplication, using Strassen's algorithm"""
-
-    # TODO: check dimensions
-
-    if len(M1) == 1:
-        return mult(M1, M2)
-
-    b = len(M1) // 2
-    A11, A12, A21, A22 = blocking(M1, b)
-    B11, B12, B21, B22 = blocking(M2, b)
-
-    K1 = mult_strassen(add(A11, A22), add(B11, B22))
-    K2 = mult_strassen(add(A21, A22), B11)
-    K3 = mult_strassen(A11, add(B12, scalar_mult(B22, -1)))
-    K4 = mult_strassen(A22, add(B21, scalar_mult(B11, -1)))
-    K5 = mult_strassen(add(A11, A12), B22)
-    K6 = mult_strassen(
-        add(A21, scalar_mult(A11, -1)), add(B11, B12)
-    )
-    K7 = mult_strassen(
-        add(A12, scalar_mult(A22, -1)), add(B21, B22)
-    )
-
-    C11 = add(K1, add(K4, add(scalar_mult(K5, -1), K7)))
-    C12 = add(K3, K5)
-    C21 = add(K2, K4)
-    C22 = add(K1, add(scalar_mult(K2, -1), add(K3, K6)))
-
-    C = []
-    for block in [[C11, C12], [C21, C22]]:
-        for row in zip(*block):
-            C.append(list(chain.from_iterable(row)))
-    return C
-
 def all_pivots_are_one(m: M) -> bool:
     pivots = list(map(lambda t : t[1] ,get_pivots(m)))
     if(next((k for k in pivots if k != 1), 1) != 1):
@@ -115,4 +151,34 @@ def below_pivots_only_zeroes(m: M) -> bool:
 def is_row_echelon_form(m: M) -> bool:
     """Function to check if Matrix m is in row_echelon_form."""
     return all_pivots_are_one(m) and below_pivots_only_zeroes(m)
+
+
+def one_step(m: M, t: list[M]) -> tuple[M, list[M]]:
+    """
+    Takes inverted list of elementary matrices, pops the first matrix
+    to apply and returns the transformed matrix as well as the remaining
+    elementary matrices.
+
+    Thus this performs a single action of Gauss elimination.
+    """
+    assert(len(t) > 0)
+    m = mult(t[0], m)
+    show(m)
+    return (m, t[1:])
+
+
+class StepByStep:
+    """A class containing matrix and a stack of elementary operations,
+    applying them one by one"""
+    def __init__(self, matrix, stack):
+        self.matrix = matrix
+        self.elementary_stack = stack
+    def __next__(self):
+        if len(self.matrix) == 0:
+            print()
+        else:
+            self.matrix = mult(self.elementary_stack[0], self.matrix)
+            show(self.matrix)
+            self.elementary_stack = self.elementary_stack[1:]
+        return self.matrix
